Create Ugwu-Daniella-Chiamaka-2024704158
The python program used the statements : if and elif with return to implement a grading system statement 
get grade function takes a score, using if and elif to determine it's boundary. When a condition is met the return statement exit the function and send back the corresponding letter grade. The else handles any scores that doesn't meet the requirement .
70-100 Grade A
60-69 Grade B
50-59 Grade C
45-49 Grade D
40-44 Grade E
39-20 Grade F

Created by: Ugwu Daniella Chiamaka 
Reg no: 2024704158
Dept: Anatomy